frontend/requirements.txt
PyQt6==6.6.1
requests==2.31.0
opencv-python==4.8.1.78
pyzbar==0.1.8
python-dotenv==1.0.0
Pillow==10.0.0
